refactor: remove Notifiers::Slack
