+++ spec/emitters/slack_spec.rb
@@ -74,14 +74,13 @@ RSpec.describe Mihari::Emitters::Slack do
     let(:mock) { double("notifier") }
 
     before do
-      allow(Mihari::Notifiers::Slack).to receive(:new).and_return(mock)
-      allow(mock).to receive(:notify)
-      allow(mock).to receive(:valid?).and_return(true)
+      allow(::Slack::Notifier).to receive(:new).and_return(mock)
+      allow(mock).to receive(:post)
     end
 
     it do
       subject.emit(title: title, description: description, artifacts: artifacts)
-      expect(mock).to have_received(:notify).once
+      expect(mock).to have_received(:post).once
     end
   end
 end
